mcp-with-law-article/requirements.txt
mcp>=0.1.0
psycopg2-binary>=2.9.0
httpx>=0.24.0
